sources: expose repoRevToNameFull as repoRevToName (#11)
Upstream nixpkgs allows for people to set a
config.fetchSourceNameDefault to determine the shape of source names
used when fetching sources. However, we want to do this always for
ekapkgs to avoid stale FODs by default. So I want to disallow usage of
the legacy "source" as a name for everything.